requirements.txt
# FastAPI and dependencies
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# Vector Database
chromadb==0.4.22

# AI/ML
anthropic==0.18.1
sentence-transformers==2.3.1
langchain==0.1.6
langchain-community==0.0.16

# Data Processing
pandas==2.2.0
numpy==1.26.3
PyPDF2==3.0.1
pdfplumber==0.10.3
PyMuPDF==1.23.8
pdf2image==1.16.3
Pillow==10.1.0
pytesseract==0.3.10

# API & Validation
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1

# Logging & Monitoring
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0